+++ alfred-telemetry-solr/alfred-telemetry-solr6/src/main/java/eu/xenit/alfred/telemetry/solr/monitoring/binder/SolrMetrics.java
@@ -23,7 +23,7 @@ public class SolrMetrics implements MeterBinder {
     @Override
     public void bindTo(MeterRegistry registry) {
         if(Util.isEnabled("METRICS_SOLR_BACKUP_ENABLED")) {
-            new SolrSnapshootMetrics(coreAdminHandler).bindTo(registry);
+            new SolrSnapshotMetrics(coreAdminHandler).bindTo(registry);
         }
         if(Util.isEnabled("METRICS_SOLR_CORESTATS_ENABLED")) {
             new SolrCoreStatsMetrics(coreAdminHandler).bindTo(registry);